Moved models and views from project to home app. This version is a successful runserver
